update labeled repo ref; make cli label helper functions as part of the library code.
