+++ snapfaas/Cargo.toml
@@ -34,7 +34,7 @@ path = "bins/sffs/main.rs"
 rand = "^0.8.5"
 ctrlc = "^3.2.0"
 reqwest = { version = "0.11", features = ["blocking"] }
-labeled = { git = "https://github.com/tan-yue/labeled", rev = "8d9fb2069e1ac7eb111f05d657af4427db600219" }
+labeled = { git = "https://github.com/alevy/labeled" }
 bytes = "1.1.0"
 byteorder = "1.2.1"
 prost = "0.9.0"
