+++ snapfaas/src/lib.rs
@@ -12,6 +12,7 @@ pub mod metrics;
 pub mod firecracker_wrapper;
 pub mod blobstore;
 pub mod labeled_fs;
+pub mod cli_utils;
 
 use std::string::String;
 use std::fs::{self, File};
